Refactor handoff validation modules
