+++ src/core/handoff_validation/__init__.py
@@ -1,0 +1,4 @@
+"""Handoff validation package."""
+from .process_flows import HandoffValidationSystem, get_handoff_validation_system
+
+__all__ = ["HandoffValidationSystem", "get_handoff_validation_system"]
