+++ src/core/handoff_validation/input_checks.py
@@ -1,0 +1,96 @@
+"""Async validation checks for handoff rules."""
+from __future__ import annotations
+
+import asyncio
+import logging
+from typing import Dict, Callable
+
+from .rules import ValidationRule, ValidationSession
+
+logger = logging.getLogger(__name__)
+
+
+async def validate_source_phase_completion(
+    rule: ValidationRule, session: ValidationSession
+) -> bool:
+    """Verify that the source phase completed all required tasks."""
+    await asyncio.sleep(0.1)
+    logger.info("🔍 Validating source phase completion for %s", session.handoff_id)
+    return True
+
+
+async def validate_target_resources(
+    rule: ValidationRule, session: ValidationSession
+) -> bool:
+    """Ensure target resources are available and accessible."""
+    await asyncio.sleep(0.1)
+    logger.info("🔍 Validating target resource availability for %s", session.handoff_id)
+    return True
+
+
+async def validate_state_consistency(
+    rule: ValidationRule, session: ValidationSession
+) -> bool:
+    """Check that system state is consistent across components."""
+    await asyncio.sleep(0.1)
+    logger.info("🔍 Validating state consistency for %s", session.handoff_id)
+    return True
+
+
+async def validate_target_agent_readiness(
+    rule: ValidationRule, session: ValidationSession
+) -> bool:
+    """Confirm the target agent is ready to receive the handoff."""
+    await asyncio.sleep(0.1)
+    logger.info("🔍 Validating target agent readiness for %s", session.handoff_id)
+    return True
+
+
+async def validate_context_transfer(
+    rule: ValidationRule, session: ValidationSession
+) -> bool:
+    """Ensure task context was transferred completely."""
+    await asyncio.sleep(0.1)
+    logger.info("🔍 Validating context transfer for %s", session.handoff_id)
+    return True
+
+
+async def validate_connection_stability(
+    rule: ValidationRule, session: ValidationSession
+) -> bool:
+    """Verify connections remain stable during handoff."""
+    await asyncio.sleep(0.1)
+    logger.info("🔍 Validating connection stability for %s", session.handoff_id)
+    return True
+
+
+async def validate_permissions(
+    rule: ValidationRule, session: ValidationSession
+) -> bool:
+    """Check that all required permissions are granted."""
+    await asyncio.sleep(0.1)
+    logger.info("🔍 Validating permissions for %s", session.handoff_id)
+    return True
+
+
+async def validate_data_integrity(
+    rule: ValidationRule, session: ValidationSession
+) -> bool:
+    """Validate that data integrity is maintained."""
+    await asyncio.sleep(0.1)
+    logger.info("🔍 Validating data integrity for %s", session.handoff_id)
+    return True
+
+
+VALIDATION_ENGINES: Dict[
+    str, Callable[[ValidationRule, ValidationSession], asyncio.Future]
+] = {
+    "source_phase_completed": validate_source_phase_completion,
+    "target_resources_available": validate_target_resources,
+    "state_consistency_verified": validate_state_consistency,
+    "target_agent_ready": validate_target_agent_readiness,
+    "context_transfer_complete": validate_context_transfer,
+    "connections_stable": validate_connection_stability,
+    "permissions_granted": validate_permissions,
+    "data_integrity_maintained": validate_data_integrity,
+}
