+++ src/core/handoff_validation_system.py
@@ -1,800 +1,4 @@
-#!/usr/bin/env python3
-"""
-Handoff Validation System - Agent Cellphone V2
-=============================================
+"""Backward compatible wrapper for the new handoff validation package."""
+from .handoff_validation import HandoffValidationSystem, get_handoff_validation_system
 
-Implements comprehensive validation and reliability testing for handoff procedures.
-This system ensures that all handoffs meet quality standards and are reliable
-under various conditions.
-
-Author: Agent-7 (QUALITY COMPLETION MANAGER)
-Contract: PHASE-003 - Smooth Handoff Procedure Implementation
-License: MIT
-"""
-
-import logging
-import time
-import asyncio
-from typing import Dict, List, Optional, Any, Callable
-from dataclasses import dataclass, field
-from enum import Enum
-from pathlib import Path
-import json
-import statistics
-
-from .base_manager import BaseManager
-
-
-class ValidationSeverity(Enum):
-    """Validation severity levels"""
-    CRITICAL = "critical"
-    HIGH = "high"
-    MEDIUM = "medium"
-    LOW = "low"
-
-
-class ValidationStatus(Enum):
-    """Validation status enumeration"""
-    PENDING = "pending"
-    IN_PROGRESS = "in_progress"
-    PASSED = "passed"
-    FAILED = "failed"
-    WARNING = "warning"
-    ERROR = "error"
-
-
-@dataclass
-class ValidationRule:
-    """Defines a validation rule"""
-    rule_id: str
-    name: str
-    description: str
-    condition: str
-    severity: ValidationSeverity
-    timeout: float = 30.0
-    retry_count: int = 3
-    dependencies: List[str] = field(default_factory=list)
-    metadata: Dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass
-class ValidationResult:
-    """Result of a validation check"""
-    rule_id: str
-    rule_name: str
-    status: ValidationStatus
-    start_time: float
-    end_time: Optional[float] = None
-    duration: Optional[float] = None
-    error_details: Optional[str] = None
-    warning_details: Optional[str] = None
-    evidence: Dict[str, Any] = field(default_factory=dict)
-    severity: ValidationSeverity = ValidationSeverity.MEDIUM
-
-
-@dataclass
-class ValidationSession:
-    """Tracks a validation session"""
-    session_id: str
-    handoff_id: str
-    procedure_id: str
-    start_time: float
-    end_time: Optional[float] = None
-    rules: List[ValidationRule] = field(default_factory=list)
-    results: List[ValidationResult] = field(default_factory=list)
-    overall_status: ValidationStatus = ValidationStatus.PENDING
-    validation_score: float = 0.0
-    critical_failures: int = 0
-    high_failures: int = 0
-    medium_failures: int = 0
-    low_failures: int = 0
-
-
-class HandoffValidationSystem(BaseManager):
-    """
-    Handoff Validation System - Quality Assurance
-    
-    Single responsibility: Validate handoff procedures and ensure reliability
-    through comprehensive testing and quality checks.
-    """
-    
-    def __init__(self, project_root: str = "."):
-        super().__init__("HandoffValidationSystem")
-        self.project_root = Path(project_root)
-        self.logger = logging.getLogger(__name__)
-        
-        # Core validation components
-        self.validation_rules: Dict[str, ValidationRule] = {}
-        self.active_sessions: Dict[str, ValidationSession] = {}
-        self.validation_history: List[ValidationSession] = []
-        
-        # Performance tracking
-        self.validation_metrics = {
-            "total_sessions": 0,
-            "successful_sessions": 0,
-            "failed_sessions": 0,
-            "average_duration": 0.0,
-            "total_duration": 0.0,
-            "rule_success_rates": {},
-            "severity_distribution": {}
-        }
-        
-        # Validation engines
-        self.validation_engines: Dict[str, Callable] = {}
-        
-        # Reliability test configurations
-        self.reliability_configs = {
-            "default_iterations": 100,
-            "default_timeout": 30.0,
-            "concurrent_limit": 10,
-            "performance_thresholds": {
-                "min_success_rate": 0.95,  # 95%
-                "max_average_duration": 5.0,  # 5 seconds
-                "max_95th_percentile": 10.0  # 10 seconds
-            }
-        }
-        
-        # Initialize the system
-        self._initialize_validation_system()
-    
-    def _initialize_validation_system(self):
-        """Initialize the validation system with default rules"""
-        self.logger.info("🚀 Initializing Handoff Validation System")
-        
-        # Load default validation rules
-        self._load_default_validation_rules()
-        
-        # Initialize validation engines
-        self._initialize_validation_engines()
-        
-        self.logger.info("✅ Handoff Validation System initialized successfully")
-    
-    def _load_default_validation_rules(self):
-        """Load default validation rules"""
-        default_rules = [
-            ValidationRule(
-                rule_id="VR001",
-                name="Phase Completion Check",
-                description="Verify that the source phase has completed all required tasks",
-                condition="source_phase_completed",
-                severity=ValidationSeverity.CRITICAL,
-                timeout=30.0,
-                retry_count=3
-            ),
-            ValidationRule(
-                rule_id="VR002",
-                name="Resource Availability Check",
-                description="Verify that target resources are available and accessible",
-                condition="target_resources_available",
-                severity=ValidationSeverity.CRITICAL,
-                timeout=45.0,
-                retry_count=2
-            ),
-            ValidationRule(
-                rule_id="VR003",
-                name="State Consistency Check",
-                description="Verify that system state is consistent across components",
-                condition="state_consistency_verified",
-                severity=ValidationSeverity.HIGH,
-                timeout=60.0,
-                retry_count=2
-            ),
-            ValidationRule(
-                rule_id="VR004",
-                name="Agent Readiness Validation",
-                description="Verify that target agent is ready to receive handoff",
-                condition="target_agent_ready",
-                severity=ValidationSeverity.CRITICAL,
-                timeout=20.0,
-                retry_count=3
-            ),
-            ValidationRule(
-                rule_id="VR005",
-                name="Context Transfer Validation",
-                description="Verify that task context was transferred completely",
-                condition="context_transfer_complete",
-                severity=ValidationSeverity.CRITICAL,
-                timeout=30.0,
-                retry_count=2
-            ),
-            ValidationRule(
-                rule_id="VR006",
-                name="Connection Stability Check",
-                description="Verify that connections are stable and reliable",
-                condition="connections_stable",
-                severity=ValidationSeverity.HIGH,
-                timeout=15.0,
-                retry_count=3
-            ),
-            ValidationRule(
-                rule_id="VR007",
-                name="Permission Validation",
-                description="Verify that all required permissions are granted",
-                condition="permissions_granted",
-                severity=ValidationSeverity.MEDIUM,
-                timeout=20.0,
-                retry_count=2
-            ),
-            ValidationRule(
-                rule_id="VR008",
-                name="Data Integrity Check",
-                description="Verify that data integrity is maintained during handoff",
-                condition="data_integrity_maintained",
-                severity=ValidationSeverity.HIGH,
-                timeout=45.0,
-                retry_count=2
-            )
-        ]
-        
-        for rule in default_rules:
-            self.validation_rules[rule.rule_id] = rule
-        
-        self.logger.info(f"📋 Loaded {len(default_rules)} default validation rules")
-    
-    def _initialize_validation_engines(self):
-        """Initialize validation engines for rule execution"""
-        self.validation_engines = {
-            "source_phase_completed": self._validate_source_phase_completion,
-            "target_resources_available": self._validate_target_resources,
-            "state_consistency_verified": self._validate_state_consistency,
-            "target_agent_ready": self._validate_target_agent_readiness,
-            "context_transfer_complete": self._validate_context_transfer,
-            "connections_stable": self._validate_connection_stability,
-            "permissions_granted": self._validate_permissions,
-            "data_integrity_maintained": self._validate_data_integrity
-        }
-    
-    def start_validation_session(self, handoff_id: str, procedure_id: str, rule_ids: Optional[List[str]] = None) -> str:
-        """
-        Start a validation session for a handoff
-        
-        Args:
-            handoff_id: ID of the handoff to validate
-            procedure_id: ID of the procedure being validated
-            rule_ids: Specific rules to validate (None for all applicable)
-            
-        Returns:
-            Session ID for tracking
-        """
-        try:
-            session_id = f"validation_{int(time.time())}_{handoff_id}"
-            
-            # Determine which rules to validate
-            if rule_ids is None:
-                rules_to_validate = list(self.validation_rules.values())
-            else:
-                rules_to_validate = [
-                    self.validation_rules[rule_id] 
-                    for rule_id in rule_ids 
-                    if rule_id in self.validation_rules
-                ]
-            
-            # Create validation session
-            session = ValidationSession(
-                session_id=session_id,
-                handoff_id=handoff_id,
-                procedure_id=procedure_id,
-                start_time=time.time(),
-                rules=rules_to_validate
-            )
-            
-            # Store session
-            self.active_sessions[session_id] = session
-            
-            # Log session start
-            self.logger.info(f"🚀 Starting validation session {session_id} for {handoff_id}")
-            
-            # Start validation
-            asyncio.create_task(self._execute_validation_session(session))
-            
-            return session_id
-            
-        except Exception as e:
-            self.logger.error(f"Failed to start validation session: {e}")
-            raise
-    
-    async def _execute_validation_session(self, session: ValidationSession):
-        """Execute a validation session"""
-        try:
-            session.overall_status = ValidationStatus.IN_PROGRESS
-            self.logger.info(f"🔄 Executing validation session {session.session_id}")
-            
-            # Execute validation for each rule
-            for rule in session.rules:
-                try:
-                    # Execute rule validation
-                    result = await self._execute_validation_rule(rule, session)
-                    session.results.append(result)
-                    
-                    # Update failure counts
-                    if result.status == ValidationStatus.FAILED:
-                        if rule.severity == ValidationSeverity.CRITICAL:
-                            session.critical_failures += 1
-                        elif rule.severity == ValidationSeverity.HIGH:
-                            session.high_failures += 1
-                        elif rule.severity == ValidationSeverity.MEDIUM:
-                            session.medium_failures += 1
-                        elif rule.severity == ValidationSeverity.LOW:
-                            session.low_failures += 1
-                    
-                    # Log result
-                    status_emoji = "✅" if result.status == ValidationStatus.PASSED else "❌"
-                    self.logger.info(f"{status_emoji} Rule {rule.rule_id} validation: {result.status.value}")
-                    
-                except Exception as e:
-                    # Create failed result
-                    result = ValidationResult(
-                        rule_id=rule.rule_id,
-                        rule_name=rule.name,
-                        status=ValidationStatus.ERROR,
-                        start_time=time.time(),
-                        end_time=time.time(),
-                        duration=0.0,
-                        error_details=str(e),
-                        severity=rule.severity
-                    )
-                    session.results.append(result)
-                    
-                    # Update failure counts
-                    if rule.severity == ValidationSeverity.CRITICAL:
-                        session.critical_failures += 1
-                    elif rule.severity == ValidationSeverity.HIGH:
-                        session.high_failures += 1
-                    elif rule.severity == ValidationSeverity.MEDIUM:
-                        session.medium_failures += 1
-                    elif rule.severity == ValidationSeverity.LOW:
-                        session.low_failures += 1
-                    
-                    self.logger.error(f"❌ Rule {rule.rule_id} validation error: {e}")
-            
-            # Calculate overall status and score
-            session.end_time = time.time()
-            session.duration = session.end_time - session.start_time
-            
-            # Determine overall status
-            if session.critical_failures > 0:
-                session.overall_status = ValidationStatus.FAILED
-            elif session.high_failures > 0:
-                session.overall_status = ValidationStatus.WARNING
-            elif session.medium_failures > 0 or session.low_failures > 0:
-                session.overall_status = ValidationStatus.WARNING
-            else:
-                session.overall_status = ValidationStatus.PASSED
-            
-            # Calculate validation score
-            total_rules = len(session.rules)
-            passed_rules = sum(1 for result in session.results if result.status == ValidationStatus.PASSED)
-            session.validation_score = passed_rules / total_rules if total_rules > 0 else 0.0
-            
-            # Log completion
-            status_emoji = "✅" if session.overall_status == ValidationStatus.PASSED else "⚠️" if session.overall_status == ValidationStatus.WARNING else "❌"
-            self.logger.info(f"{status_emoji} Validation session {session.session_id} completed: {session.overall_status.value} (Score: {session.validation_score:.2%})")
-            
-            # Update metrics
-            self._update_validation_metrics(session)
-            
-        except Exception as e:
-            session.overall_status = ValidationStatus.ERROR
-            session.end_time = time.time()
-            session.duration = session.end_time - session.start_time
-            self.logger.error(f"❌ Validation session {session.session_id} failed: {e}")
-            
-            # Update metrics
-            self._update_validation_metrics(session)
-        
-        finally:
-            # Move to history
-            self.validation_history.append(session)
-            if session.session_id in self.active_sessions:
-                del self.active_sessions[session_id]
-    
-    async def _execute_validation_rule(self, rule: ValidationRule, session: ValidationSession) -> ValidationResult:
-        """Execute a single validation rule"""
-        try:
-            start_time = time.time()
-            
-            # Check if validation engine exists
-            if rule.condition not in self.validation_engines:
-                raise ValueError(f"Unknown validation condition: {rule.condition}")
-            
-            # Execute validation with timeout and retries
-            validation_success = False
-            error_details = None
-            
-            for attempt in range(rule.retry_count + 1):
-                try:
-                    validation_result = await asyncio.wait_for(
-                        self.validation_engines[rule.condition](rule, session),
-                        timeout=rule.timeout
-                    )
-                    
-                    if validation_result:
-                        validation_success = True
-                        break
-                    
-                except asyncio.TimeoutError:
-                    error_details = f"Validation timed out after {rule.timeout}s (attempt {attempt + 1})"
-                    if attempt == rule.retry_count:
-                        break
-                except Exception as e:
-                    error_details = f"Validation error (attempt {attempt + 1}): {e}"
-                    if attempt == rule.retry_count:
-                        break
-            
-            end_time = time.time()
-            duration = end_time - start_time
-            
-            # Determine result status
-            if validation_success:
-                status = ValidationStatus.PASSED
-            elif rule.severity == ValidationSeverity.CRITICAL:
-                status = ValidationStatus.FAILED
-            else:
-                status = ValidationStatus.WARNING
-            
-            # Create validation result
-            result = ValidationResult(
-                rule_id=rule.rule_id,
-                rule_name=rule.name,
-                status=status,
-                start_time=start_time,
-                end_time=end_time,
-                duration=duration,
-                error_details=error_details,
-                severity=rule.severity
-            )
-            
-            return result
-            
-        except Exception as e:
-            end_time = time.time()
-            duration = end_time - start_time
-            
-            # Create error result
-            result = ValidationResult(
-                rule_id=rule.rule_id,
-                rule_name=rule.name,
-                status=ValidationStatus.ERROR,
-                start_time=start_time,
-                end_time=end_time,
-                duration=duration,
-                error_details=str(e),
-                severity=rule.severity
-            )
-            
-            return result
-    
-    # Validation engine implementations
-    async def _validate_source_phase_completion(self, rule: ValidationRule, session: ValidationSession) -> bool:
-        """Validate that the source phase has completed"""
-        try:
-            # Simulate phase completion validation
-            await asyncio.sleep(0.1)
-            
-            # Check if phase completion criteria are met
-            # This would typically involve checking:
-            # - All required tasks completed
-            # - All deliverables delivered
-            # - All quality gates passed
-            # - All dependencies resolved
-            
-            self.logger.info(f"🔍 Validating source phase completion for {session.handoff_id}")
-            
-            # Simulate successful validation
-            return True
-            
-        except Exception as e:
-            self.logger.error(f"Source phase completion validation failed: {e}")
-            return False
-    
-    async def _validate_target_resources(self, rule: ValidationRule, session: ValidationSession) -> bool:
-        """Validate that target resources are available"""
-        try:
-            # Simulate resource availability check
-            await asyncio.sleep(0.1)
-            
-            # Check if target resources are available
-            # This would typically involve checking:
-            # - Resource availability
-            # - Resource accessibility
-            # - Resource permissions
-            # - Resource capacity
-            
-            self.logger.info(f"🔍 Validating target resources for {session.handoff_id}")
-            
-            # Simulate successful validation
-            return True
-            
-        except Exception as e:
-            self.logger.error(f"Target resources validation failed: {e}")
-            return False
-    
-    async def _validate_state_consistency(self, rule: ValidationRule, session: ValidationSession) -> bool:
-        """Validate that system state is consistent"""
-        try:
-            # Simulate state consistency check
-            await asyncio.sleep(0.1)
-            
-            # Check if system state is consistent
-            # This would typically involve checking:
-            # - State synchronization
-            # - Data consistency
-            # - Configuration consistency
-            # - Cache consistency
-            
-            self.logger.info(f"🔍 Validating state consistency for {session.handoff_id}")
-            
-            # Simulate successful validation
-            return True
-            
-        except Exception as e:
-            self.logger.error(f"State consistency validation failed: {e}")
-            return False
-    
-    async def _validate_target_agent_readiness(self, rule: ValidationRule, session: ValidationSession) -> bool:
-        """Validate that target agent is ready"""
-        try:
-            # Simulate agent readiness check
-            await asyncio.sleep(0.1)
-            
-            # Check if target agent is ready
-            # This would typically involve checking:
-            # - Agent online status
-            # - Agent capabilities
-            # - Agent resources
-            # - Agent workload
-            
-            self.logger.info(f"🔍 Validating target agent readiness for {session.handoff_id}")
-            
-            # Simulate successful validation
-            return True
-            
-        except Exception as e:
-            self.logger.error(f"Target agent readiness validation failed: {e}")
-            return False
-    
-    async def _validate_context_transfer(self, rule: ValidationRule, session: ValidationSession) -> bool:
-        """Validate that context transfer is complete"""
-        try:
-            # Simulate context transfer validation
-            await asyncio.sleep(0.1)
-            
-            # Check if context transfer is complete
-            # This would typically involve checking:
-            # - Context data completeness
-            # - Reference updates
-            # - Dependency resolution
-            # - State preservation
-            
-            self.logger.info(f"🔍 Validating context transfer for {session.handoff_id}")
-            
-            # Simulate successful validation
-            return True
-            
-        except Exception as e:
-            self.logger.error(f"Context transfer validation failed: {e}")
-            return False
-    
-    async def _validate_connection_stability(self, rule: ValidationRule, session: ValidationSession) -> bool:
-        """Validate that connections are stable"""
-        try:
-            # Simulate connection stability check
-            await asyncio.sleep(0.1)
-            
-            # Check if connections are stable
-            # This would typically involve checking:
-            # - Connection health
-            # - Latency measurements
-            # - Packet loss
-            # - Connection reliability
-            
-            self.logger.info(f"🔍 Validating connection stability for {session.handoff_id}")
-            
-            # Simulate successful validation
-            return True
-            
-        except Exception as e:
-            self.logger.error(f"Connection stability validation failed: {e}")
-            return False
-    
-    async def _validate_permissions(self, rule: ValidationRule, session: ValidationSession) -> bool:
-        """Validate that permissions are granted"""
-        try:
-            # Simulate permission validation
-            await asyncio.sleep(0.1)
-            
-            # Check if permissions are granted
-            # This would typically involve checking:
-            # - Access permissions
-            # - Resource permissions
-            # - Operation permissions
-            # - Security policies
-            
-            self.logger.info(f"🔍 Validating permissions for {session.handoff_id}")
-            
-            # Simulate successful validation
-            return True
-            
-        except Exception as e:
-            self.logger.error(f"Permissions validation failed: {e}")
-            return False
-    
-    async def _validate_data_integrity(self, rule: ValidationRule, session: ValidationSession) -> bool:
-        """Validate that data integrity is maintained"""
-        try:
-            # Simulate data integrity check
-            await asyncio.sleep(0.1)
-            
-            # Check if data integrity is maintained
-            # This would typically involve checking:
-            # - Data consistency
-            # - Checksum validation
-            # - Transaction integrity
-            # - Backup verification
-            
-            self.logger.info(f"🔍 Validating data integrity for {session.handoff_id}")
-            
-            # Simulate successful validation
-            return True
-            
-        except Exception as e:
-            self.logger.error(f"Data integrity validation failed: {e}")
-            return False
-    
-    def _update_validation_metrics(self, session: ValidationSession):
-        """Update validation performance metrics"""
-        try:
-            self.validation_metrics["total_sessions"] += 1
-            
-            if session.overall_status == ValidationStatus.PASSED:
-                self.validation_metrics["successful_sessions"] += 1
-            else:
-                self.validation_metrics["failed_sessions"] += 1
-            
-            # Calculate duration
-            if session.duration:
-                self.validation_metrics["total_duration"] += session.duration
-                self.validation_metrics["average_duration"] = (
-                    self.validation_metrics["total_duration"] / self.validation_metrics["total_sessions"]
-                )
-            
-            # Update rule success rates
-            for result in session.results:
-                rule_id = result.rule_id
-                if rule_id not in self.validation_metrics["rule_success_rates"]:
-                    self.validation_metrics["rule_success_rates"][rule_id] = {
-                        "total": 0,
-                        "passed": 0,
-                        "failed": 0
-                    }
-                
-                self.validation_metrics["rule_success_rates"][rule_id]["total"] += 1
-                if result.status == ValidationStatus.PASSED:
-                    self.validation_metrics["rule_success_rates"][rule_id]["passed"] += 1
-                else:
-                    self.validation_metrics["rule_success_rates"][rule_id]["failed"] += 1
-            
-            # Update severity distribution
-            for result in session.results:
-                severity = result.severity.value
-                if severity not in self.validation_metrics["severity_distribution"]:
-                    self.validation_metrics["severity_distribution"][severity] = {
-                        "total": 0,
-                        "passed": 0,
-                        "failed": 0
-                    }
-                
-                self.validation_metrics["severity_distribution"][severity]["total"] += 1
-                if result.status == ValidationStatus.PASSED:
-                    self.validation_metrics["severity_distribution"][severity]["passed"] += 1
-                else:
-                    self.validation_metrics["severity_distribution"][severity]["failed"] += 1
-                
-        except Exception as e:
-            self.logger.error(f"Failed to update validation metrics: {e}")
-    
-    # Public interface methods
-    def get_validation_status(self, session_id: str) -> Optional[Dict[str, Any]]:
-        """Get the status of a validation session"""
-        try:
-            if session_id in self.active_sessions:
-                session = self.active_sessions[session_id]
-                return {
-                    "session_id": session.session_id,
-                    "handoff_id": session.handoff_id,
-                    "procedure_id": session.procedure_id,
-                    "status": session.overall_status.value,
-                    "progress": len(session.results) / len(session.rules) if session.rules else 0.0,
-                    "start_time": session.start_time,
-                    "rules_total": len(session.rules),
-                    "rules_completed": len(session.results)
-                }
-            else:
-                # Check history
-                for validation in self.validation_history:
-                    if validation.session_id == session_id:
-                        return {
-                            "session_id": validation.session_id,
-                            "handoff_id": validation.handoff_id,
-                            "procedure_id": validation.procedure_id,
-                            "status": validation.overall_status.value,
-                            "validation_score": validation.validation_score,
-                            "start_time": validation.start_time,
-                            "end_time": validation.end_time,
-                            "duration": validation.duration,
-                            "critical_failures": validation.critical_failures,
-                            "high_failures": validation.high_failures,
-                            "medium_failures": validation.medium_failures,
-                            "low_failures": validation.low_failures
-                        }
-                
-                return None
-                
-        except Exception as e:
-            self.logger.error(f"Failed to get validation status: {e}")
-            return None
-    
-    def get_system_status(self) -> Dict[str, Any]:
-        """Get overall system status and metrics"""
-        try:
-            return {
-                "system_status": "operational",
-                "active_sessions": len(self.active_sessions),
-                "total_sessions": self.validation_metrics["total_sessions"],
-                "success_rate": (
-                    self.validation_metrics["successful_sessions"] / self.validation_metrics["total_sessions"]
-                    if self.validation_metrics["total_sessions"] > 0 else 0.0
-                ),
-                "average_duration": self.validation_metrics["average_duration"],
-                "available_rules": list(self.validation_rules.keys()),
-                "rule_success_rates": self.validation_metrics["rule_success_rates"],
-                "severity_distribution": self.validation_metrics["severity_distribution"]
-            }
-        except Exception as e:
-            self.logger.error(f"Failed to get system status: {e}")
-            return {"error": str(e)}
-    
-    def add_validation_rule(self, rule: ValidationRule) -> bool:
-        """Add a new validation rule"""
-        try:
-            if rule.rule_id in self.validation_rules:
-                self.logger.warning(f"Rule {rule.rule_id} already exists, overwriting")
-            
-            self.validation_rules[rule.rule_id] = rule
-            self.logger.info(f"✅ Added validation rule: {rule.rule_id}")
-            return True
-            
-        except Exception as e:
-            self.logger.error(f"Failed to add validation rule: {e}")
-            return False
-    
-    def remove_validation_rule(self, rule_id: str) -> bool:
-        """Remove a validation rule"""
-        try:
-            if rule_id not in self.validation_rules:
-                self.logger.warning(f"Rule {rule_id} not found")
-                return False
-            
-            # Check if rule is in use
-            for session in self.active_sessions.values():
-                for rule in session.rules:
-                    if rule.rule_id == rule_id:
-                        self.logger.error(f"Cannot remove rule {rule_id} - in use by session {session.session_id}")
-                        return False
-            
-            del self.validation_rules[rule_id]
-            self.logger.info(f"✅ Removed validation rule: {rule_id}")
-            return True
-            
-        except Exception as e:
-            self.logger.error(f"Failed to remove validation rule: {e}")
-            return False
-
-
-# Global instance for system-wide access
-handoff_validation_system = HandoffValidationSystem()
-
-
-def get_handoff_validation_system() -> HandoffValidationSystem:
-    """Get the global handoff validation system instance"""
-    return handoff_validation_system
+__all__ = ["HandoffValidationSystem", "get_handoff_validation_system"]
